backend/requirements.txt
fastapi
uvicorn
motor
python-dotenv
python-jose
passlib
pydantic[email]